+++ src/main/java/kitchenpos/deliveryorder/application/port/in/DeliveryOrderUseCase.java
@@ -1,10 +1,5 @@
 package kitchenpos.deliveryorder.application.port.in;
 
-import java.util.List;
-import java.util.UUID;
-
-import kitchenpos.deliveryorder.domain.DeliveryOrder;
-
 public interface DeliveryOrderUseCase {
 
 }
