+++ src/main/java/kitchenpos/deliveryorder/domain/OrderLineItem.java
@@ -1,5 +1,4 @@
 package kitchenpos.deliveryorder.domain;
 
-public class DeliveryOrderLineItem {
-
+public class OrderLineItem {
 }
