--- src/main/java/kitchenpos/eatinTable/domain/EatInTable.java
@@ -1,5 +1,0 @@
-package kitchenpos.eatinTable.domain;
-
-public class EatInTable {
-
-}
