+++ src/main/java/kitchenpos/eatinorder/domain/EatInTable.java
@@ -1,0 +1,5 @@
+package kitchenpos.eatinorder.domain;
+
+public class EatInTable {
+
+}
