+++ src/main/java/kitchenpos/eatinorder/domain/OrderLineItem.java
@@ -1,0 +1,4 @@
+package kitchenpos.eatinorder.domain;
+
+public class OrderLineItem {
+}
