+++ src/main/java/kitchenpos/menu/adapter/in/MenuGroupRestController.java
@@ -1,17 +1,7 @@
 package kitchenpos.menu.adapter.in;
 
-import java.net.URI;
-import java.util.List;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import kitchenpos.menu.application.MenuGroupService;
-import kitchenpos.menu.domain.MenuGroup;
 
 @RequestMapping("/api/menu-groups")
 @RestController
