+++ src/main/java/kitchenpos/menu/adapter/in/ProductRestController.java
@@ -1,9 +1,10 @@
-package kitchenpos.eatinTable.adapter.in;
+package kitchenpos.menu.adapter.in;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@RequestMapping("/api/eatin-tables")
+@RequestMapping("/api/products")
 @RestController
-public class EatinTableRestController {
+public class ProductRestController {
+
 }
