+++ src/main/java/kitchenpos/menu/application/port/in/MenuUseCase.java
@@ -1,10 +1,5 @@
 package kitchenpos.menu.application.port.in;
 
-import java.util.List;
-import java.util.UUID;
-
-import kitchenpos.menu.domain.Menu;
-
 public interface MenuUseCase {
 
 }
