+++ src/main/java/kitchenpos/menu/application/port/in/ProductUseCase.java
@@ -1,5 +1,5 @@
 package kitchenpos.menu.application.port.in;
 
-public interface MenuGroupUseCase {
+public interface ProductUseCase {
 
 }
