+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -1,21 +1,5 @@
 package kitchenpos.menu.domain;
 
-import java.util.UUID;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.ForeignKey;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-import javax.persistence.Transient;
-
-import kitchenpos.product.domain.Product;
-
-
 public class MenuProduct {
 
 }
