--- src/main/java/kitchenpos/product/adapter/in/ProductRestController.java
@@ -1,23 +1,0 @@
-package kitchenpos.product.adapter.in;
-
-import java.net.URI;
-import java.util.List;
-import java.util.UUID;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import kitchenpos.product.application.ProductService;
-import kitchenpos.product.domain.Product;
-
-@RequestMapping("/api/products")
-@RestController
-public class ProductRestController {
-
-}
