--- src/main/java/kitchenpos/product/application/port/in/ProductUseCase.java
@@ -1,10 +1,0 @@
-package kitchenpos.product.application.port.in;
-
-import java.util.List;
-import java.util.UUID;
-
-import kitchenpos.product.domain.Product;
-
-public interface ProductUseCase {
-
-}
