+++ src/main/java/kitchenpos/takeoutorder/application/port/in/TakeoutOrderUseCase.java
@@ -1,10 +1,5 @@
 package kitchenpos.takeoutorder.application.port.in;
 
-import java.util.List;
-import java.util.UUID;
-
-import kitchenpos.takeoutorder.domain.TakeoutOrder;
-
 public interface TakeoutOrderUseCase {
 
 }
