+++ src/main/java/kitchenpos/takeoutorder/domain/OrderLineItem.java
@@ -1,0 +1,4 @@
+package kitchenpos.takeoutorder.domain;
+
+public class OrderLineItem {
+}
