--- src/main/java/kitchenpos/takeoutorder/domain/TakeoutOrderLineItem.java
@@ -1,20 +1,0 @@
-package kitchenpos.takeoutorder.domain;
-
-import java.math.BigDecimal;
-import java.util.UUID;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.ForeignKey;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-import javax.persistence.Transient;
-
-import kitchenpos.menu.domain.Menu;
-
-public class TakeoutOrderLineItem {
-}
